# openapi.yml
openapi: 3.0.0
info:
  description: Mozilla's IAM group API
  version: 1.0.0-oas3
  title: Groups
  contact:
    email: fiji@mozilla.com
  license:
    name: MPL 2.0
    url: https://www.mozilla.org/media/MPL/2.0/index.txt
paths:
...
  "/groups/api/v1/terms/{groupName}":
    get:
      summary: group terms
      description: get the terms of condition for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "200":
          description: text of the groups terms in markdown
          content:
            application/json:
              schema:
                type: string
                example: some terms
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    delete:
      summary: delete group terms
      description: delete the terms of condition for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "201":
          description: terms deleted
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    post:
      summary: create group terms
      description: create terms of condition for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: some terms
      responses:
        "201":
          description: terms created
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    put:
      summary: update group terms
      description: update terms of condition for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: some terms
      responses:
        "201":
          description: terms updated
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
...
  "/groups/api/v1/members/{groupName}/{memberUuid}/renew":
    post:
      summary: renew a member
      description: renew a member for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: path
          name: memberUuid
          description: pass in the uuid of the member to renew
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group_expiration:
                  description: expiration period in days
                  type: integer
                  example: 180
      responses:
        "201":
          description: member renewed
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
  "/groups/api/v1/members/{groupName}/{memberUuid}":
    delete:
      summary: delete a member
      description: delete a member of a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: path
          name: memberUuid
          description: pass in the uuid of the member to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: member deleted
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
  "/groups/api/v1/members/{groupName}":
    get:
      summary: paginated members
      description: |
        get paginated members of a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: query
          name: r
          description: filter results based on role
          required: false
          schema:
            type: string
            enum: [Any, Member, Curator]
            default: Any
        - in: query
          name: q
          description: search query to filter results
          required: false
          schema:
            type: string
        - in: query
          name: s
          description: number of results
          required: false
          schema:
            type: integer
            default: 20
        - in: query
          name: "n"
          description: next results token
          required: false
          schema:
            type: integer
        - in: query
          name: by
          description: sort by
          required: false
          schema:
            type: string
            enum: [None, ExpirationAsc, ExpirationDesc, RoleAsc, RoleDesc]
            default: None
      responses:
        "200":
          description: a page of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/DisplayMember"
                  next:
                    type: integer
                    example: 20
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
...
  "/groups/api/v1/sudo/member/{groupName}":
    post:
      summary: add a member
      description: add a member to a group (admins only)
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: path
          name: memberUuid
          description: pass in the uuid of the member to add
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_uuid:
                  type: string
                  format: uuid
                  example: 34a33a8e-d035-4051-acac-497a79235e73
                group_expiration:
                  description: expiration period in days
                  type: integer
                  example: 180
      responses:
        "201":
          description: member added
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
servers:
  - url: http://localhost:8085
components:
  schemas:
    GenericError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Group name exists
...
    DisplayMember:
      type: object
      properties:
        user_uuid:
          type: string
          format: uuid
          example: 34a33a8e-d035-4051-acac-497a79235e73
        picture:
          type: string
          nullable: true
          example: https://picture.api.dev.sso.allizom.org/avatar/get/id/MzI0ZjAwMzA5Yzc1ZGYxYmIzYjdjMGE1NWZmMDc3OGY3NGJjYjAzMjc0NDVhZmI2YWViODBmMzIzOGNhYzM3NyNzdGFmZiMxNTcyODYyNzM3.png
        first_name:
          type: string
          nullable: true
          example: Hans
        last_name:
          type: string
          nullable: true
          example: Knall
        username:
          type: string
          example: hans
        email:
          type: string
          format: email
          nullable: true
          example: hans@knall.org
        isStaff:
          type: boolean
          example: true
        since:
          type: string
          format: datetime
          nullable: true
          example: 2014-11-28T12:45:59Z
        expiration:
          type: string
          format: datetime
          nullable: true
          example: 2014-11-28T12:45:59Z
        role:
          type: string
          enum:
            - Admin
            - Curator
            - Member
        added_by:
          $ref: "#/components/schemas/AddedBy"
...
    AddedBy:
      type: object
      properties:
        user_uuid:
          type: string
          example: 34a33a8e-d035-4051-acac-497a79235e73
        first_name:
          type: string
          nullable: true
          example: Hans
        last_name:
          type: string
          nullable: true
          example: Knall
        username:
          type: string
          nullable: true
          example: hans
        email:
          type: string
          format: email
          nullable: true
          example: hans@knall.org
...